Moved RockSack class to its own set of files.
